move code to utility
